Implement JWT auth, onboarding, and in-memory user management
